+++ src/form/b_checkbox.rs
@@ -11,11 +11,12 @@ pub fn BCheckbox(
     label: &'static str,
     #[prop(optional, into)] name: Option<&'static str>,
     #[prop(default = "true".into(), into)] value: MaybeSignal<String>,
+    #[prop(optional, into)] is_checked: MaybeSignal<bool>,
     #[prop(optional, into)] on_change: Option<EventFn>,
     #[prop(attrs, optional)] attributes: Vec<(&'static str, Attribute)>,
 ) -> impl IntoView {
     let mut b_input =
-        view! { <input node_ref=node_ref id=id name=name type="checkbox" value=value/> }
+        view! { <input node_ref=node_ref id=id name=name type="checkbox" value=value checked=is_checked/> }
             .optional_event(ev::change, on_change.map(EventFn::into_inner));
 
     for (attr_name, attr_value) in attributes {