+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Optional prop `is_checked` on checkbox.
+
 ## [0.0.0-alpha.11] - 2024-04-13
 
 ### Changed